% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_specific_mcmc_sampler.R
\name{day_specific_fosr}
\alias{day_specific_fosr}
\title{MCMC Sampling Algorithm for the Function-on-Scalars Regression Model
with extra level of hierarchy (for day-specific effects in Actigraphy example)}
\usage{
day_specific_fosr(Y, tau, U, X = NULL, K = NULL, nsave = 1000,
  nburn = 1000, nskip = 3, mcmc_params = list("beta", "fk", "alpha",
  "sigma_e", "sigma_g"), computeDIC = TRUE)
}
\arguments{
\item{Y}{the \code{n x m} data observation matrix, where \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{the \code{m x d} matrix of coordinates of observation points}

\item{U}{the \code{n} vector identifying the subject.}

\item{X}{the \code{n x p} matrix of predictors; if NULL, only include an intercept}

\item{K}{the number of factors; if NULL, use SVD-based proportion of variability explained}

\item{nsave}{number of MCMC iterations to record}

\item{nburn}{number of MCMC iterations to discard (burin-in)}

\item{nskip}{number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{mcmc_params}{named list of parameters for which we store the MCMC output;
must be one or more of
\itemize{
\item "beta" (factors)
\item "fk" (loading curves)
\item "alpha" (regression coefficients)
\item "mu_k" (intercept term for factor k)
\item "sigma_e" (observation error SD)
\item "sigma_g" (random effects SD)
\item "Yhat" (fitted values)
\item "Psi" (factor loadings)
\item "gamma" (random error terms)
}}

\item{computeDIC}{logical; if TRUE, compute the deviance information criterion \code{DIC}
and the effective number of parameters \code{p_d}}
}
\value{
A named list of the \code{nsave} MCMC samples for the parameters named in \code{mcmc_params}
}
\description{
Let n = sum_i=1,v u_i where u_i is the number of observations for observation i
}
\note{
If \code{nm} is large, then storing all posterior samples for \code{Yhat}, which is \code{nsave x n x M},  may be inefficient
}
