% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_post_alpha_tilde}
\alias{get_post_alpha_tilde}
\title{Compute the posterior distrubution for the regression coefficient functions}
\usage{
get_post_alpha_tilde(post_fk, post_alpha_j)
}
\arguments{
\item{post_fk}{\code{Nsims x m x K} matrix of posterior draws of the loading curve matrix}

\item{post_alpha_j}{\code{Nsims x K} matrix of posterior draws of the regression coefficient factors}
}
\value{
\code{Nsims x m} matrix of posterior draws of the regression coefficient function
}
\description{
Given posterior samples for the loading curves \code{fk} and the
regression coefficient factors \code{alpha_j} for a predictor \code{j},
compute the posterior distribution of the corresponding regression coefficient function.
}
