% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_factors}
\alias{plot_factors}
\title{Plot the factors}
\usage{
plot_factors(post_beta, subj = NULL)
}
\arguments{
\item{post_beta}{the \code{Nsims x n x K} array of \code{Nsims} draws from the posterior
distribution of the \code{n x K} matrix of factors, \code{beta}}

\item{subj}{\code{n x 1} vector of subject IDs or labels}
}
\description{
Plot posterior mean of the factors together with the simultaneous and pointwise
95\% credible bands.
}
