% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot the Bayesian curve fitted values}
\usage{
plot_fitted(y, mu, postY, y_true = NULL, t01 = NULL,
  include_joint_bands = FALSE)
}
\arguments{
\item{y}{the \code{n x 1} vector of time series observations}

\item{mu}{the \code{n x 1} vector of fitted values, i.e., posterior expectation of the mean}

\item{postY}{the \code{nsims x n} matrix of posterior draws from which to compute intervals}

\item{y_true}{the \code{n x 1} vector of points along the true curve}

\item{t01}{the observation points; if NULL, assume \code{n} equally spaced points from 0 to 1}

\item{include_joint_bands}{logical; if TRUE, compute simultaneous credible bands}
}
\description{
Plot the curve posterior means with posterior credible intervals (pointwise and joint),
the observed data, and true curves (if known)
}
\examples{
# FIXME

}
