% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simulate_fosr}
\alias{simulate_fosr}
\title{Simulate a function-on-scalar regression model}
\usage{
simulate_fosr(
  n = 100,
  m = 50,
  RSNR = 5,
  K_true = 4,
  p_0 = 1000,
  p_1 = 5,
  sparse_factors = TRUE,
  corr = 0,
  perc_missing = 0,
  X = NULL
)
}
\arguments{
\item{n}{number of observed curves (i.e., number of subjects)}

\item{m}{total number of observation points (i.e., points along the curve)}

\item{RSNR}{root signal-to-noise ratio}

\item{K_true}{rank of the model (i.e., number of basis functions used for the functional data simulations)}

\item{p_0}{number of true zero regression coefficients}

\item{p_1}{number of true nonzero regression coefficients}

\item{sparse_factors}{logical; if TRUE, then for each nonzero predictor j,
sample a subset of k=1:K_true factors to be nonzero#'}

\item{corr}{correlation parameter for predictors}

\item{perc_missing}{percentage of missing data (between 0 and 1); default is zero}

\item{X}{the design matrix to use. By default, \code{X} is NULL. When \code{X} is NULL, a design matrix
is created. If \code{X} is specified, then params \code{p_0}, \code{n} and}
}
\value{
a list containing the following:
\itemize{
\item \code{Y}: the simulated \code{n x m} functional data matrix
\item \code{X}: the simulated \code{n x p} design matrix
\item \code{tau}: the \code{m}-dimensional vector of observation points
\item \code{Y_true}: the true \code{n x m} functional data matrix (w/o noise)
\item \code{alpha_tilde_true} the true \code{m x p} matrix of regression coefficient functions
\item \code{alpha_arr_true} the true \code{K_true x p} matrix of regression coefficient factors
\item \code{Beta_true} the true \code{n x K_true} matrix of factors
\item \code{F_true} the true \code{m x K_true} matrix of basis (loading curve) functions
\item \code{sigma_true} the true observation error standard deviation
}
}
\description{
Simulate data from a function-on-scalar regression model, allowing for
subject-specific random effects. The predictors are multivariate normal with
mean zero and covariance \code{corr^abs(j1-j2)} for correlation parameter \code{corr}
between predictors \code{j1} and \code{j2}.
More predictors than observations (p > n) is allowed.
}
\note{
The basis functions (or loading curves) are orthonormalized polynomials,
so large values of \code{K_true} are not recommended.
}
\examples{
# Example: simulate FOSR
sim_data = simulate_fosr(n = 100, m = 20, p_0 = 100, p_1 = 5)
Y = sim_data$Y; X = sim_data$X; tau = sim_data$tau

}
